==> scripts/1_EXTRACTING-CORPORA.ipynb <==


**Make sure that you are connected to the kernel associated with our virtual environment . Go to `Kernel` -> `Change kernel` and choose `pia_venv`.**

import pandas as pd
import requests

# Load the main dataset of ancient Greek texts

# LAGT (v1.0.1) dataset from public folder
resp = requests.get("https://sciencedata.dk/public/164bd4b21e23faeb39221fb2ad03f16a/AGT_lemmatized_20201211.json")
LAGT = pd.DataFrame(resp.json())

# get some overview of the shape of the data
LAGT.head(5)

def flat_lemmata(sentences):
    lemmata = [word for sent in sentences for word in sent]
    return lemmata
LAGT["lemmata"] = LAGT["lemmatized_sentences"].apply(flat_lemmata)

LAGT["lemmata_wordcount"] = LAGT["lemmata"].apply(lambda x: len(x))
LAGT["lemmata_wordcount"].sum() # previously we had 13925726, then 13713183, finally 14756899

# Data Extraction and Overview



### how many documents we have
len(LAGT)

### let's identify our main author of interest

c_hippocraticum = LAGT[LAGT["author_id"].str.startswith("tlg0627")]
len(c_hippocraticum) ### old value: 53

len(c_hippocraticum[c_hippocraticum["filename"].str.contains("perseus")])

c_hippocraticum["n_sentences"].sum()

c_hippocraticum["wordcount"].sum() # old value: 443514, then 333446, 333443

c_hippocraticum["lemmata_wordcount"].sum()

# for comparison, we also generate corpus aristotelicum
c_aristotelicum = LAGT[LAGT["author_id"].str.startswith("tlg0086")]
len(c_aristotelicum) ### originally we had 27

c_aristotelicum["wordcount"].sum() # 857024

# However in the case of Aristotle, we are interested only in a subselection of works associated with his name
# we coded them manually in the "include" column
c_aristotelicum_coded = pd.read_csv("../data/c_aristotelicum_OVERVIEW.csv")

### we can check whether the new and old version have the same order
print(c_aristotelicum_coded["doc_id"].tolist())
print(c_aristotelicum["doc_id"].tolist())

# 1) add the "include?" column from the old overview to our new dataset 
c_aristotelicum["include?"] = c_aristotelicum_coded["include?"].tolist()
# 2) use the "include?" column for filtering the new dataset
# 3) drop "include?" column 
c_aristotelicum = c_aristotelicum[c_aristotelicum["include?"]=="y"].drop("include?", axis=1)
c_aristotelicum.head()

len(c_aristotelicum)

c_aristotelicum["wordcount"].sum()

# Export the corpora

c_aristotelicum.to_json("../data/c_aristotelicum.json")
c_hippocraticum.to_json("../data/c_hippocraticum.json")
