==> scripts/1_LEMMATIZATION.ipynb <==
# Requirements

import requests
import pandas as pd
import numpy as np

# our own package for reading the data
#!pip install sddk
import sddk

from anda import gr
# OR uncomment the following:
# script_url = "https://raw.githubusercontent.com/sdam-au/anda_py/master/anda/gr.py"
# exec(requests.get(script_url).content)

# to use the latest version (for development)
from importlib import reload
reload(gr)

conf = sddk.configure()

# Usage

If you have only one word, you can either return all possible lemmata associated with it separated by `/` (`return_all_unique_lemmata()`), or you can return only the first of them (`return_first_lemma()`). In  case of all lemmatization functions, there is a set of optional additional arguments:
* `all_lemmata=False`
* `filter_by_postag=None` (try `filter_by_postag=["n","a","v"]` to get nouns, adjectives and verbs
* `involve_unknown=False`, i.e. return (True) or omit (False) words in cases where no lemma was found)

gr.return_first_lemma("ἐπιστήμην")

gr.return_all_unique_lemmata("ἐπιστήμην")

You can also lemmatize an extensive string, consisting from a number of words, like this one:

string = "Πρότασις μὲν οὖν ἐστὶ λόγος καταφατικὸς ἢ ἀποφατικὸς τινὸς κατά τινος. Οὗτος δὲ ἢ καθόλου ἢ ἐν μέρει ἢ ἀδιόριστος. Λέγω δὲ καθόλου μὲν τὸ παντὶ ἢ μηδενὶ ὑπάρχειν, ἐν μέρει δὲ τὸ τινὶ ἢ μὴ τινὶ ἢ μὴ παντὶ ὑπάρχειν, ἀδιόριστον δὲ τὸ ὑπάρχειν ἢ μὴ ὑπάρχειν ἄνευ τοῦ καθόλου, ἢ κατὰ μέρος, οἷον τὸ τῶν ἐναντίων εἶναι τὴν αὐτὴν ἐπιστήμην ἢ τὸ τὴν ἡδονὴν μὴ εἶναι ἀγαθόν."

Then you run:

lemmatized_string = gr.lemmatize_string(string)
print(lemmatized_string)

The results above are still quite poor. Perhaps there is a problem with encoding...

You can also lemmatize the string and at the same time split it into sentences:

lemmatized_sentences = gr.get_lemmatized_sentences(string)
print(lemmatized_sentences)

lemmatized_sentences = gr.get_lemmatized_sentences(string, all_lemmata=True, filter_by_postag=["n","a","v"], involve_unknown=False)
print(lemmatized_sentences)

gr.return_all_unique_translations("ὑπάρχειν", filter_by_postag=None, involve_unknown=False)

gr.lemma_translator("λόγος")

string_tokenized = gr.tokenize_string(string)

gr.morphological_analysis(string)[1:4]

word = "ζῆν"
list_of_tokens = gr.return_list_of_tokens(word)
list_of_tokens

gr.lemmatize_string("ζῆν", all_lemmata=False, filter_by_postag=["n","a","v"], involve_unknown=True)

# Import corpora

c_aristotelicum = sddk.read_file("c_aristotelicum.json", "df", "31b393e2afe1ee96ce81869c7efe18cb")
c_hippocraticum = sddk.read_file("c_hippocraticum.json", "df", "31b393e2afe1ee96ce81869c7efe18cb")

# Testing lemmatization with Nicomachean Ethics

This two corpora already contain one version of lemmatization, but here we will improve it

EN_df = c_aristotelicum[c_aristotelicum["doc_id"]=="tlg0086.tlg010"]
EN_df

# Nicomachean ethics as a string"
EN_df_string = EN_df["string"].tolist()[0]
EN_df_string[:500]

# lemmatization without any parametrization  
print(gr.lemmatize_string(EN_df_string[:500]))

# lemmatization with all lemmata
print(gr.lemmatize_string(EN_df_string[:500], all_lemmata=True))

# lemmatization with all lemmata, but only nouns, adjectives and verbs
# prefer noun and adjective where noun, adjective and verb is available
# prefer noun where noun and adjective is available
print(gr.lemmatize_string(EN_df_string[:500], all_lemmata=True, filter_by_postag=["n","a","v"], involve_unknown=True))

EN_df_string[:500]

print(gr.lemmatize_string(EN_df_string[:500], all_lemmata=False, filter_by_postag=["n","a","v"], involve_unknown=True))

print(gr.get_lemmatized_sentences(EN_df_string[:1000], all_lemmata=False, filter_by_postag=["n","a","v"], involve_unknown=True))

print(gr.get_lemmatized_sentences(EN_df_string[:1000], all_lemmata=False, filter_by_postag=["n","a","v"], involve_unknown=False))

list_of_tokens = gr.return_list_of_tokens("καλῶς")
list_of_tokens

### simple test of how preference works
list_of_tokens = gr.return_list_of_tokens("καλῶς")
token_letters = set([token["p"][0] for token in list_of_tokens])
if (("n" or "a") in token_letters) and ("v" in token_letters):
    print("met")
    list_of_tokens = [token for token in list_of_tokens if token["p"][0] in ["a", "n"]]
# prefer nouns over adjectives
if "a" and "n" in token_letters:
    list_of_tokens = [token for token in list_of_tokens if token["p"][0] == "n"]
list_of_tokens

# Main Lemmatization

%%time
c_aristotelicum["lemmata"] = c_aristotelicum.apply(lambda row: gr.lemmatize_string(row["string"], all_lemmata=False, filter_by_postag=["n","a","v"], involve_unknown=True), axis=1)
c_hippocraticum["lemmata"] = c_hippocraticum.apply(lambda row: gr.lemmatize_string(row["string"], all_lemmata=False, filter_by_postag=["n","a","v"], involve_unknown=True), axis=1)

%%time
c_aristotelicum["lemmatized_sentences"] = c_aristotelicum.apply(lambda row: gr.get_lemmatized_sentences(row["string"], all_lemmata=False, filter_by_postag=["n","a","v"], involve_unknown=True), axis=1)
c_hippocraticum["lemmatized_sentences"] = c_hippocraticum.apply(lambda row: gr.get_lemmatized_sentences(row["string"], all_lemmata=False, filter_by_postag=["n","a","v"], involve_unknown=True), axis=1)

c_aristotelicum.head(5)

#conf = sddk.configure()
sddk.write_file("public_data/PIA/c_aristotelicum.json", c_aristotelicum, conf)
sddk.write_file("public_data/PIA/c_hippocraticum.json", c_hippocraticum, conf)
