==> scripts/5_EMBEDDINGS.ipynb <==
### PREREQUISTIES
### (many used only in one notebook...)

import os
import pandas as pd
import numpy as np
import logging ### to monitor the code
from bs4 import BeautifulSoup
import pickle
import math
import random
import sys
import csv
import unicodedata
import requests
import re

import nltk
from nltk.collocations import *

import matplotlib.pyplot as plt
import seaborn as sns

from scipy import stats

import xml.etree.cElementTree as ET
from urllib.request import urlopen

import gspread
from gspread_dataframe import get_as_dataframe, set_with_dataframe
#from google.colab import auth
#from oauth2client.client import GoogleCredentials
from google.oauth2 import service_account # based on google-auth library
import sddk

# gensim parts
from gensim import corpora
from gensim import models
from gensim.models import CoherenceModel, LdaModel, LsiModel, HdpModel
from gensim.models.wrappers import LdaMallet

ἢν

#!pip install anda
#from anda import gr ### the import takes substantial time, since it import a +600MB file containing ancient Greek dictionary

### not neccessary for reading the data, just for exporting them to sciencedata.dk
conf = sddk.configure()

# to access gsheet, you need Google Service Account key json file
# I have mine located in my personal space on sciencedata.dk, so I read it from there:

# (1) read the file and parse its content
file_data = conf[0].get(conf[1] + "ServiceAccountsKey.json").json()
# (2) transform the content into crendentials object
credentials = service_account.Credentials.from_service_account_info(file_data)
# (3) specify your usage of the credentials
scoped_credentials = credentials.with_scopes(['https://spreadsheets.google.com/feeds', 'https://www.googleapis.com/auth/drive'])
# (4) use the constrained credentials for authentication of gspread package
gc = gspread.Client(auth=scoped_credentials)
# (5) establish connection with spreadsheets specified by their url
PIA_data = gc.open_by_url("https://docs.google.com/spreadsheets/d/1KxOx7Be9fj3lDcEPgQhQ-Iqcn9p367-MMD6RMXe8rks/edit?usp=sharing")
PIA_overview = gc.open_by_url("https://docs.google.com/spreadsheets/d/1e94wyelg6dftQ4zxbq1xvwxWAI-BhcYXtclDW-YTnrw/edit?usp=sharing")

publicfolder = "31b393e2afe1ee96ce81869c7efe18cb"
c_hippocraticum = sddk.read_file("c_hippocraticum_enriched.json", "df", publicfolder)

words_to_remove = ["ἢν", "ὁκόταν"]

c_hippocraticum["lemmatized_sentences_repl"] = c_hippocraticum["lemmatized_sentences_repl"].apply(lambda sentences_list:  [[word for word in sentence if word not in words_to_remove] for sentence in sentences_list])

words_to_replace = {
    "ἔχις" : "ἔχω"}

for key in words_to_replace.keys():
    c_hippocraticum["lemmatized_sentences_repl"] = c_hippocraticum["lemmatized_sentences_repl"].apply(lambda sentences_list: [[corrections[key] if x == key else x for x in sentence] for sentence in sentences_list])

# Corpora

c_hippocraticum.columns

def get_flat_sentences(series):
    sentences_list = [sent for doc in series.tolist() for sent in doc]
    return sentences_list

sentences_list = get_flat_sentences(c_hippocraticum["lemmatized_sentences_repl"])

flat_sentences = []
for sentence in sentences_list:
    flat_sentences.extend(sentence)

print(sentences_list[:10])

c_hippocraticum["lemmata_repl"].apply(lambda x: len(x)).sum()

subcorpora = {}
for cat in c_hippocraticum["work_cat_linka"].unique():
    subcorpora[cat] = get_flat_sentences(c_hippocraticum[c_hippocraticum["work_cat_linka"]==cat]["lemmatized_sentences_repl"])

for key in subcorpora.keys():
    print(key, len(subcorpora[key]))

#create dictionary for our list of sentences
dictionary = corpora.Dictionary(sentences_list)

there is a number of methods to be applied upon a dictionary object:
see: https://radimrehurek.com/gensim/corpora/dictionary.html

First of all, you can inspect it as a standard dictionary object


# uncomment below to print the whole dictionary
# dictionary

# as such, it is organized by ids
dict(list(dictionary.items())[:10])

# but you can access it reveresly by applying token2id method
dict(list(dictionary.token2id.items())[:10])

# collection_frequencies
dictionary.cfs[5] # how many instances of word with id 5 (= "θάνατος")

dictionary.dfs[5] # how many documents contain the word with id 5 (= "θάνατος")

dictionary.num_pos

len(dictionary.keys())

sent = sentences_list[20]
sent

# doc2bow = document to (term, tf) tuples, i.e. bag-of-words
dictionary.doc2bow(sent)

# build bag-of-word corpus for all sentences 
corpus = [dictionary.doc2bow(sent) for sent in sentences_list]

# look at first five sentences:
print(corpus[:5])

# tfidf
from gensim import models
tfidf = models.TfidfModel(corpus)

for sentence in tfidf[corpus][:5]:
    print(sentence)

# LSI model

num_topics = 10
lsimodel = LsiModel(corpus=corpus, num_topics=num_topics, id2word=dictionary)

lsimodel.show_topic(0)

topics_dict = {} 
for n in range(num_topics):
    topics_dict["topic " + str(n)] = [tup[0] for tup in lsimodel.show_topic(n, topn=10)]
lsi_topics = pd.DataFrame(topics_dict)
lsi_topics

lsimodel.print_topic(4)

lsimodel_df = pd.DataFrame(lsimodel.get_topics(), columns=dictionary.values()).T

 # get first topic, i.e. sort words by largest 
lsimodel_df.sort_values(4, ascending=False).head(20)

lsimodel_df_sorted = lsimodel_df.sort_values(4, ascending=False)  
len(lsimodel_df_sorted.loc[:"χρόνος"])

words = ['λύπ*', 'ἄλγ*', 'ὀδύν*', 'πόνο*']

len(lsimodel_df)

word_positions = {}
for word in words:
    positions = []
    for n in range(num_topics):
        lsimodel_df_sorted = lsimodel_df.sort_values(n, ascending=False)  
        pos = len(lsimodel_df_sorted.loc[:word])
        positions.append(pos)
    word_positions[word] = positions
word_positions

word_positions = {}
for word in words:
    positions = []
    for n in range(num_topics):
        lsimodel_df_sorted = lsimodel_df.sort_values(n, ascending=False)  
        pos = len(lsimodel_df_sorted.loc[:word])
        positions.append(len(lsimodel_df.loc[:word]))
    word_positions[word] = positions
word_positions

columns = ["pos in top " + str(n) for n in range(num_topics)]

word_positions_df = pd.DataFrame(word_positions).T
word_positions_df.columns = columns
word_positions_df

lsimodel_df[0].index

# HDP model

hdpmodel = HdpModel(corpus=corpus, id2word=dictionary)

hdpmodel_df = pd.DataFrame(hdpmodel.get_topics(), columns=dictionary.values()).T

hdpmodel
